+++ CHANGES.rst
@@ -1,8 +1,12 @@
 Products.MeetingPROVHainaut Changelog
 =====================================
 
-4.1 (2019-06-25)
+4.1 (unreleased)
 ----------------
+- Manage zprovhainaut install profile
+- Create and configure finances and compta advices
+- Adapt comptabilite Workflow to remove relevant states
+- Override MeetingItem.mayEvaluateCompleteness as only finances/comptabilite precontrollers may evaluate it
 - Added new field for grouped items on a slip number
 
 4.0 (2018-10-25)